Relocate token gates under roles and permissions
