+++ components/settings/roles/token-gates/TokenGates.tsx
@@ -19,7 +19,7 @@ import { getLitChainFromChainId } from 'lib/token-gates';
 import { useSnackbar } from 'hooks/useSnackbar';
 import useSWR from 'swr';
 import ConfirmDeleteModal from 'components/common/Modal/ConfirmDeleteModal';
-import Legend from '../Legend';
+import Legend from '../../Legend';
 import TokenGatesTable from './TokenGatesTable';
 
 // Example: https://github.com/LIT-Protocol/lit-js-sdk/blob/9b956c0f399493ae2d98b20503c5a0825e0b923c/build/manual_tests.html
